Rentals module created
