+++ app.rb
@@ -11,6 +11,7 @@ require './rental'
 require './classroom'
 require './student'
 require './teacher'
+require './rentals_handler'
 
 class App
   attr_reader :books, :people, :rentals
@@ -20,7 +21,7 @@ class App
     @people = []
     @rentals = []
   end
-
+  include Rentals
   # is teacher or student
   def teacher_or_student(type)
     case type
@@ -88,25 +89,7 @@ class App
     puts
   end
 
-  def create_rental
-    puts 'Select a book from the following list by number:'
-    list_books
-    book_id = gets.chomp
-    book = @books[book_id.to_i - 1]
-    puts 'Select a person from the following list by number:'
-    list_people
-    person_id = gets.chomp
-    person = @people[person_id.to_i - 1]
-    print 'Date:  '
-    date = gets.chomp
-    new_rental = Rental.new(book, person, date)
-    new_rental.person.id = person_id
-    @rentals.push(new_rental)
-    puts
-    puts "Rental for #{new_rental.person.id} created successfully."
-    puts
-    puts
-  end
+
 
   def list_books
     @books.each_with_index { |book, index| puts "#{index + 1}) Title: \"#{book.title}\", Author: #{book.author}" }
